pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "Production-grade domain-specific RAG system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.4.0",
    "langchain-chroma>=0.2.0",
    "sentence-transformers>=3.0.0",
    "rank-bm25>=0.2.2",
    "pdfplumber>=0.11.0",
    "tiktoken>=0.9.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "datasets>=3.0.0",
    "fpdf2>=2.8.7",
    "chromadb>=1.5.5",
    "langchain-ollama>=1.0.1",
    "langchain-core>=1.2.19",
    "fastapi>=0.135.1",
    "uvicorn>=0.42.0",
    "python-multipart>=0.0.22",
    "streamlit>=1.55.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
eval = ["ragas>=0.2.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
